Fix PRIME lead route reference
