+++ lib/features/prime/prime_lead_screen.dart
@@ -1,0 +1,187 @@
+import 'package:flutter/material.dart';
+
+class PrimeLeadScreen extends StatefulWidget {
+  const PrimeLeadScreen({super.key});
+
+  @override
+  State<PrimeLeadScreen> createState() => _PrimeLeadScreenState();
+}
+
+class _PrimeLeadScreenState extends State<PrimeLeadScreen> {
+  final _formKey = GlobalKey<FormState>();
+  final _nameCtrl = TextEditingController();
+  final _phoneCtrl = TextEditingController();
+  final _goalCtrl = TextEditingController();
+  final _messageCtrl = TextEditingController(
+    text: 'Quiero activar PRIME COLOSO lo antes posible.',
+  );
+
+  bool _sending = false;
+
+  @override
+  void dispose() {
+    _nameCtrl.dispose();
+    _phoneCtrl.dispose();
+    _goalCtrl.dispose();
+    _messageCtrl.dispose();
+    super.dispose();
+  }
+
+  Future<void> _submit() async {
+    if (!_formKey.currentState!.validate()) {
+      return;
+    }
+    setState(() => _sending = true);
+    await Future<void>.delayed(const Duration(milliseconds: 600));
+    if (!mounted) return;
+    setState(() => _sending = false);
+    await showDialog<void>(
+      context: context,
+      builder: (ctx) => AlertDialog(
+        backgroundColor: const Color(0xFF111111),
+        title: const Text(
+          'Solicitud enviada',
+          style: TextStyle(color: Colors.white, fontWeight: FontWeight.w800),
+        ),
+        content: const Text(
+          'Tu coach recibirá esta información para continuar el cierre de tu acceso PRIME COLOSO. En breve te contactaremos '
+          'para finalizar el pago y activar tus beneficios.',
+          style: TextStyle(color: Colors.white70, height: 1.4),
+        ),
+        actions: [
+          TextButton(
+            onPressed: () => Navigator.of(ctx).pop(),
+            style: TextButton.styleFrom(foregroundColor: Colors.white),
+            child: const Text('Entendido'),
+          ),
+        ],
+      ),
+    );
+    if (!mounted) return;
+    Navigator.of(context).pop();
+  }
+
+  InputDecoration _decoration(String label) {
+    return InputDecoration(
+      labelText: label,
+      labelStyle: const TextStyle(color: Colors.white70, fontWeight: FontWeight.w600),
+      filled: true,
+      fillColor: const Color(0xFF1B1B1B),
+      enabledBorder: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(18),
+        borderSide: const BorderSide(color: Color(0xFF2C2C2C)),
+      ),
+      focusedBorder: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(18),
+        borderSide: const BorderSide(color: Color(0xFFD0202A), width: 1.6),
+      ),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final textTheme = Theme.of(context).textTheme;
+
+    return Scaffold(
+      backgroundColor: Colors.black,
+      appBar: AppBar(
+        backgroundColor: Colors.black,
+        elevation: 0,
+        iconTheme: const IconThemeData(color: Colors.white),
+        title: const Text(
+          'Activa PRIME COLOSO',
+          style: TextStyle(color: Colors.white, fontWeight: FontWeight.w800),
+        ),
+      ),
+      body: SafeArea(
+        child: SingleChildScrollView(
+          padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 24),
+          child: Form(
+            key: _formKey,
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  'Comparte tus datos y un coach iniciará el cierre de tu membresía PRIME COLOSO.',
+                  style: textTheme.bodyLarge?.copyWith(
+                    color: Colors.white,
+                    fontWeight: FontWeight.w600,
+                  ),
+                ),
+                const SizedBox(height: 24),
+                TextFormField(
+                  controller: _nameCtrl,
+                  style: const TextStyle(color: Colors.white),
+                  decoration: _decoration('Nombre completo'),
+                  validator: (value) =>
+                      value == null || value.trim().isEmpty ? 'Ingresa tu nombre' : null,
+                ),
+                const SizedBox(height: 18),
+                TextFormField(
+                  controller: _phoneCtrl,
+                  style: const TextStyle(color: Colors.white),
+                  decoration: _decoration('WhatsApp o teléfono de contacto'),
+                  keyboardType: TextInputType.phone,
+                  validator: (value) =>
+                      value == null || value.trim().isEmpty ? 'Necesitamos un contacto' : null,
+                ),
+                const SizedBox(height: 18),
+                TextFormField(
+                  controller: _goalCtrl,
+                  style: const TextStyle(color: Colors.white),
+                  decoration: _decoration('¿Cuál es tu objetivo Coloso?'),
+                ),
+                const SizedBox(height: 18),
+                TextFormField(
+                  controller: _messageCtrl,
+                  style: const TextStyle(color: Colors.white),
+                  decoration: _decoration('Mensaje para tu coach'),
+                  maxLines: 5,
+                ),
+                const SizedBox(height: 32),
+                SizedBox(
+                  width: double.infinity,
+                  child: FilledButton(
+                    onPressed: _sending ? null : _submit,
+                    style: FilledButton.styleFrom(
+                      backgroundColor: const Color(0xFFD0202A),
+                      padding: const EdgeInsets.symmetric(vertical: 18),
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(18),
+                      ),
+                    ),
+                    child: _sending
+                        ? const SizedBox(
+                            height: 20,
+                            width: 20,
+                            child: CircularProgressIndicator(
+                              strokeWidth: 2,
+                              valueColor:
+                                  AlwaysStoppedAnimation<Color>(Colors.white),
+                            ),
+                          )
+                        : const Text(
+                            'Enviar solicitud',
+                            style: TextStyle(
+                              fontWeight: FontWeight.w800,
+                              fontSize: 16,
+                            ),
+                          ),
+                  ),
+                ),
+                const SizedBox(height: 18),
+                Text(
+                  'Después de validar tu pago manualmente activaremos tu cuenta como PRIME COLOSO para que desbloquees todos los accesos.',
+                  style: textTheme.bodySmall?.copyWith(
+                    color: Colors.white54,
+                    height: 1.4,
+                  ),
+                ),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
